error fixes

+++ cms/controller/modals/edit.php
@@ -20,36 +20,38 @@ if ($_SERVER['REQUEST_METHOD'] == 'POST') {
 
     $modal = new Modal();
     $id = $_POST['id'];
-    echo json_encode($_FILES);
-    echo json_encode($_POST);
-    die();
+    // echo json_encode($_FILES);
+    // echo json_encode($_POST);
+    // die();
     $res = $modal->edit_modal($id, $_POST);
     $data = array();
-    for($i = 0; $i < count($_POST['title_eng']); $i++) {
-        if($_POST['title_eng'][$i] != '') {
-            $data[] = [
-                'modal_id' => $id,
-                'title_eng' => $_POST['title_eng'][$i],
-                'title_ar' => $_POST['title_ar'][$i],
-                'text_eng' => $_POST['text_eng'][$i],
-                'text_ar' => $_POST['text_ar'][$i],
-            ];
-            if($_FILES['media']['tmp_name'][$i] == '' && $_POST['old_media'][$i] != '') {
-                $data[$i]['media'] = $_POST['old_media'][$i];
-                $data[$i]['type'] = $_POST['old_type'][$i];
-                $data[$i]['filetype'] = $_POST['old_filetype'][$i];
-            } else if ($_FILES['media']['tmp_name'][$i] != '') {
-                $targetDir = $items_config['modal_media_path'];
-                if (!file_exists($targetDir)) {
-                    @mkdir($targetDir);
-                }
-                $file = $_FILES['media']['tmp_name'][$i];
-                $fileName = time().'_'.$_FILES['media']['name'][$i]; 
-                $targetFile = $targetDir . '/' . $fileName;
-                if (move_uploaded_file($file, $targetFile)) {
-                    $data[$i]['media'] = $fileName;
-                    $data[$i]['type'] = explode('/',$_FILES['media']['type'][$i])[0];
-                    $data[$i]['filetype'] = $_FILES['media']['type'][$i];
+    if(array_key_exists('title_eng', $_POST)) {
+        for($i = 0; $i < count($_POST['title_eng']); $i++) {
+            if($_POST['title_eng'][$i] != '') {
+                $data[] = [
+                    'modal_id' => $id,
+                    'title_eng' => $_POST['title_eng'][$i],
+                    'title_ar' => $_POST['title_ar'][$i],
+                    'text_eng' => $_POST['text_eng'][$i],
+                    'text_ar' => $_POST['text_ar'][$i],
+                ];
+                if($_FILES['media']['tmp_name'][$i] == '' && $_POST['old_media'][$i] != '') {
+                    $data[$i]['media'] = $_POST['old_media'][$i];
+                    $data[$i]['type'] = $_POST['old_type'][$i];
+                    $data[$i]['filetype'] = $_POST['old_filetype'][$i];
+                } else if ($_FILES['media']['tmp_name'][$i] != '') {
+                    $targetDir = $items_config['modal_media_path'];
+                    if (!file_exists($targetDir)) {
+                        @mkdir($targetDir);
+                    }
+                    $file = $_FILES['media']['tmp_name'][$i];
+                    $fileName = time().'_'.$_FILES['media']['name'][$i]; 
+                    $targetFile = $targetDir . '/' . $fileName;
+                    if (move_uploaded_file($file, $targetFile)) {
+                        $data[$i]['media'] = $fileName;
+                        $data[$i]['type'] = explode('/',$_FILES['media']['type'][$i])[0];
+                        $data[$i]['filetype'] = $_FILES['media']['type'][$i];
+                    }
                 }
             }
         }
@@ -57,7 +59,10 @@ if ($_SERVER['REQUEST_METHOD'] == 'POST') {
     // echo json_encode($data);
     // die();
     $modal->remove_prev_modal_items($id);
-    $res = $modal->add_modal_item($data);
+    if($data) {
+        $res = $modal->add_modal_item($data);
+    }
+    
 
     if ($res) {
         $_SESSION['success'] = 'Modal Updated Successfully';
@@ -77,8 +82,10 @@ if($_SERVER['REQUEST_METHOD'] == 'GET') {
 	$timeline_obj = new Timeline();
     $timelines = $timeline_obj->get_timeline_items();
     $medias = array();
-    foreach ($item_media as $media) {
-        array_push($medias, $media);
+    if($item_media) {
+        foreach ($item_media as $media) {
+            array_push($medias, $media);
+        }
     }
     $modal['items'] = $medias;
 }
